Moving KubernetesObjectsProvider interface to kubernetes-node to avoid a cyclic dependency
