+++ plugins/kubernetes-backend/src/plugin.ts
@@ -21,13 +21,11 @@ import {
 } from '@backstage/backend-plugin-api';
 import { catalogServiceRef } from '@backstage/plugin-catalog-node/alpha';
 
-import {
-  KubernetesBuilder,
-  KubernetesObjectsProvider,
-} from '@backstage/plugin-kubernetes-backend';
+import { KubernetesBuilder } from '@backstage/plugin-kubernetes-backend';
 import {
   KubernetesObjectsProviderExtensionPoint,
   kubernetesObjectsProviderExtensionPoint,
+  KubernetesObjectsProvider,
 } from '@backstage/plugin-kubernetes-node';
 
 class ObjectsProvider implements KubernetesObjectsProviderExtensionPoint {
