+++ plugins/kubernetes-backend/src/types/index.ts
@@ -15,3 +15,9 @@
  */
 
 export * from './types';
+
+export type {
+  CustomResourcesByEntity,
+  KubernetesObjectsByEntity,
+  KubernetesObjectsProvider,
+} from '@backstage/plugin-kubernetes-node';
