+++ plugins/kubernetes-backend/src/types/types.ts
@@ -21,9 +21,7 @@ import type {
   CustomResourceMatcher,
   FetchResponse,
   KubernetesFetchError,
-  KubernetesRequestAuth,
   KubernetesRequestBody,
-  ObjectsByEntityResponse,
 } from '@backstage/plugin-kubernetes-common';
 import { Config } from '@backstage/config';
 import { KubernetesCredential } from '../auth/types';
@@ -229,33 +227,3 @@ export interface KubernetesObjectsProviderOptions {
  * @public
  */
 export type ObjectsByEntityRequest = KubernetesRequestBody;
-
-/**
- *
- * @public
- */
-export interface KubernetesObjectsByEntity {
-  entity: Entity;
-  auth: KubernetesRequestAuth;
-}
-
-/**
- *
- * @public
- */
-export interface CustomResourcesByEntity extends KubernetesObjectsByEntity {
-  customResources: CustomResourceMatcher[];
-}
-
-/**
- *
- * @public
- */
-export interface KubernetesObjectsProvider {
-  getKubernetesObjectsByEntity(
-    kubernetesObjectsByEntity: KubernetesObjectsByEntity,
-  ): Promise<ObjectsByEntityResponse>;
-  getCustomResourcesByEntity(
-    customResourcesByEntity: CustomResourcesByEntity,
-  ): Promise<ObjectsByEntityResponse>;
-}
