+++ plugins/kubernetes-node/api-report.md
@@ -3,13 +3,43 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { CustomResourceMatcher } from '@backstage/plugin-kubernetes-common';
+import { Entity } from '@backstage/catalog-model';
 import { ExtensionPoint } from '@backstage/backend-plugin-api';
-import { KubernetesObjectsProvider } from '@backstage/plugin-kubernetes-backend';
+import { KubernetesObjectsProvider as KubernetesObjectsProvider_2 } from '@backstage/plugin-kubernetes-node';
+import { KubernetesRequestAuth } from '@backstage/plugin-kubernetes-common';
+import { ObjectsByEntityResponse } from '@backstage/plugin-kubernetes-common';
+
+// @public (undocumented)
+export interface CustomResourcesByEntity extends KubernetesObjectsByEntity {
+  // (undocumented)
+  customResources: CustomResourceMatcher[];
+}
+
+// @public (undocumented)
+export interface KubernetesObjectsByEntity {
+  // (undocumented)
+  auth: KubernetesRequestAuth;
+  // (undocumented)
+  entity: Entity;
+}
+
+// @public (undocumented)
+export interface KubernetesObjectsProvider {
+  // (undocumented)
+  getCustomResourcesByEntity(
+    customResourcesByEntity: CustomResourcesByEntity,
+  ): Promise<ObjectsByEntityResponse>;
+  // (undocumented)
+  getKubernetesObjectsByEntity(
+    kubernetesObjectsByEntity: KubernetesObjectsByEntity,
+  ): Promise<ObjectsByEntityResponse>;
+}
 
 // @public
 export interface KubernetesObjectsProviderExtensionPoint {
   // (undocumented)
-  addObjectsProvider(provider: KubernetesObjectsProvider): void;
+  addObjectsProvider(provider: KubernetesObjectsProvider_2): void;
 }
 
 // @public
