+++ plugins/kubernetes-node/src/types/index.ts
@@ -14,24 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * Node.js library for the kubernetes plugin.
- *
- * @packageDocumentation
- */
-
-// In this package you might for example export functions that
-// help other plugins or modules interact with your plugin.
-
-/**
- * Node.js library for the kubernetes plugin.
- *
- * @packageDocumentation
- */
-
-export {
-  kubernetesObjectsProviderExtensionPoint,
-  type KubernetesObjectsProviderExtensionPoint,
-} from './extensions';
-
 export * from './types';
