+++ plugins/kubernetes-node/src/types/types.ts
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2023 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { Entity } from '@backstage/catalog-model';
+import {
+  CustomResourceMatcher,
+  KubernetesRequestAuth,
+  ObjectsByEntityResponse,
+} from '@backstage/plugin-kubernetes-common';
+
+/**
+ *
+ * @public
+ */
+
+export interface KubernetesObjectsProvider {
+  getKubernetesObjectsByEntity(
+    kubernetesObjectsByEntity: KubernetesObjectsByEntity,
+  ): Promise<ObjectsByEntityResponse>;
+  getCustomResourcesByEntity(
+    customResourcesByEntity: CustomResourcesByEntity,
+  ): Promise<ObjectsByEntityResponse>;
+}
+
+/**
+ *
+ * @public
+ */
+
+export interface KubernetesObjectsByEntity {
+  entity: Entity;
+  auth: KubernetesRequestAuth;
+}
+/**
+ *
+ * @public
+ */
+
+export interface CustomResourcesByEntity extends KubernetesObjectsByEntity {
+  customResources: CustomResourceMatcher[];
+}
